requirements.txt
# Python dependencies
python-docx
PyPDF2
pytesseract
opencv-python
matplotlib
python-pptx
fpdf
